Move private types/functions out of public header.
Improves compilation times of gfx and any code that includes gfx.h.
